READ THIS COMMIT!!!
I've added some important verbiage around the account recovery mechanism, and its security considerations.

I will not mince words: This gives instance operators two ways to impersonate users. Any use of the BurnDown mechanism should be scrutinized to hell.

However, I believe it strikes the most reasonable balance possible.

I'm sure someone will, in the future, point to this as a reason to distrust me or this project. I want to make it clear that the intent is to move the needle for encryption online.

I'm not building a Signal competitor. I don't want to build a Signal competitor. I just want E2EE to be table stakes, and this PKD project will help make it possible.

Skepticism is healthy, but don't freak out because a more paranoid project would choose differently.