Better guidelines on previous Merkle root associations
The entire security goal of this feature is to prevent pre-computation attacks against the plaintext commitment, without relying on a random value (e.g., salt) that a malicious Directory could manipulate to trick a client into accepting the wrong plaintext for a given committed ciphertext. We don't need unique values or even "always the latest" value.