Don't allow encrypted BurnDown
We assume the instance moderation team is responsible for issuing BurnDown messages. If we allowed them to be encrypted by the client, they could send a BurnDown for any user since the PKD has NO way of knowing who's an admin or not. It's not our job to know, either!