Stub out algorithm selection
This updates the specification text to make it easier for algorithm suites to be specified that use alternative cryptographic primitives.

For example, if someone (for whatever reason) wanted a CNSA 2.0 and/or FIPS-140 compatible alternative to v1, they could use a NIST SP800-108 KDF with ML-DSA-44 for signatures.

(I personally don't see any value in FIPS for this project, but stranger things have happened.)