Remove TODO.
We were going to bootstrap a list of trusted PKD instances from the client and server, and rely on the intersection of both. But that's mildly annoying.

We also considered a "Vouch" mechanism for which users could state "I have seen this public key directory before and trust its contents", but that's too Web of Trust-y.

I'll revisit this.